Same as for steim1.
Hello,

I was using your Java implementation of the Steim1 decompression function and discovered some discrepancies to the official libmseed library.

I attached a 1kb data block with a 64 byte header from a Reftek digitizer.

If I read that block with the attached Java program, I can get up to 21 samples out of it where the last 5 samples are all the same.

If I read it with the libmseed library with e.g. Obspy with the attached Python script I get only (maximum) 17 samples (also other methods using libmseed yield that number of samples).

The problem seems to be in the extractSamples method where you switch over the curNibble and add samples if it is 0. The libmseed library skips the case 0 completely and includes the first samples separately. If the case 0 is only handled if the offset is 0 AND i < 3 it seems to fix the problem. However I am not sure if this introduces other problems. The following code would include the fix:

switch (currNibble) {
case 0:
// System.out.println("0 means header info");
// only include header info if offset is 0
if (offset == 0 && i < 3) {
temp[currNum++] = Utility.bytesToInt(
bytes[offset + (i * 4)],
bytes[offset + (i * 4) + 1], bytes[offset + (i * 4)
+ 2], bytes[offset + (i * 4) + 3],
swapBytes);
}
break;


Another possibility could be that my sample does not fulfill the Steim1 standard and your code is actually right. I don't know the standard well enough to know that.

Best Regards,
Benjamin Sick